LIE_CONNECTED/requirements.txt
flask
flask-cors
joblib
numpy
librosa
scikit-learn
praat-parselmouth